Refactor host page (#4488)
